// src/app/service/model/UserEditDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface UserEditDTO {
    /**
     * Email of User.
     */
    email?: string;

    links?: Array<models.Link>;

    /**
     * Login name of the User.
     */
    login?: string;

    /**
     * Display nickname of the User.
     */
    nickName?: string;

    /**
     * Password hash of the User.
     */
    password?: string;

    /**
     * Registration date of User in long.
     */
    registered?: Date;

    /**
     * Roles of the User (authorizations).
     */
    roles?: Array<UserEditDTO.RolesEnum>;

    /**
     * Unique identifier.
     */
    userId?: number;

}
export namespace UserEditDTO {
    export enum RolesEnum {
        ADMIN = <any> 'ADMIN',
        AUTHOR = <any> 'AUTHOR',
        LOCKED = <any> 'LOCKED'
    }
}